move all button components
