+++ packages/~/common/ui/package.json
@@ -5,21 +5,18 @@
   "type": "module",
   "exports": {
     ".": {
-      "types": "./src/index.ts",
       "default": "./src/index.ts"
     },
     "./*": {
-      "types": "./src/*.ts",
       "default": "./src/*.ts"
     }
   },
-  "main": "./src/index.ts",
-  "types": "./src/index.ts",
   "dependencies": {},
   "devDependencies": {
     "@happy-dom/global-registrator": "^13.3.8",
     "@tsconfig/bun": "1.0.1",
     "bun-types": "1.0.25",
+    "hono": "4.0.2",
     "htmx.org": "1.9.10",
     "hyperscript.org": "0.9.12"
   },
