+++ packages/~/common/ui/src/button/components/copy.tsx
@@ -2,7 +2,7 @@
 
 import type { Child } from "hono/jsx";
 import type { VariantProps } from "tailwind-variants";
-import { button } from ".";
+import { button } from "..";
 
 //
