+++ src/moderations/:id/02.tsx
@@ -4,7 +4,7 @@ import { api_ref } from ":api_ref";
 import type { Organization } from ":database:moncomptepro";
 import { app_hc } from ":hc";
 import { ORGANISATION_EVENTS } from ":organizations/services/event";
-import { button } from ":ui/button";
+import { button } from "@~/common.ui/button/index";
 import { useContext } from "hono/jsx";
 import { ModerationPage_Context } from "./page";
