+++ src/moderations/:id/About_Organisation.tsx
@@ -1,7 +1,7 @@
 //
 
 import { app_hc } from ":hc";
-import { button } from ":ui/button";
+import { button } from "@~/common.ui/button/index";
 import { useContext } from "hono/jsx";
 import { ModerationPage_Context } from "./page";
