+++ src/moderations/page.tsx
@@ -7,8 +7,8 @@ import { z_coerce_boolean } from ":common/z.coerce.boolean";
 import type { Moderation, Organization } from ":database:moncomptepro";
 import type { moncomptepro_pg_Context } from ":database:moncomptepro/middleware";
 import { app_hc } from ":hc";
-import { button } from ":ui/button";
 import { row } from ":ui/table";
+import { button } from "@~/common.ui/button/index";
 import { useRequestContext } from "hono/jsx-renderer";
 import { z } from "zod";
 import { moderation_type_to_emoji } from "./moderation_type_to_emoji";
